requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
